more warnings fixed

+++ Source/PC/Graphics_PC.cpp
@@ -340,7 +340,7 @@ void cGraphics_PC::video_Draw_Sprite() {
 	Fodder->word_42074 = 160 - Fodder->word_4206C;
 	Fodder->word_4206C >>= 1;
 	
-	Fodder->word_42076 = mImage->GetWidth() - (Fodder->word_4206C*4);
+	Fodder->word_42076 = (uint16) (mImage->GetWidth() - (Fodder->word_4206C*4));
 
 	di += Plane;
 	for( int16 dx = Fodder->word_4206E; dx > 0; --dx ) {
@@ -663,7 +663,7 @@ void cGraphics_PC::Recruit_Draw_HomeAway( ) {
 	mFodder->sub_13C1C( 0x0E, 0x0A, 0, 0x9B );
 	
 	std::string Home = tool_StripLeadingZero(tool_NumToString( mFodder->word_397AE ));
-	mFodder->sub_16B55( 0x0D, (int16) 0x9A - (Home.length() * 0x0C), 0x0A, Home );
+	mFodder->sub_16B55( 0x0D, (int16) (0x9A - (Home.length() * 0x0C)), 0x0A, Home );
 
 	std::string Away = tool_StripLeadingZero(tool_NumToString( mFodder->word_397AC ));
 	mFodder->sub_16B55( 0x0D, 0xAA, 0x0A, Away );
